deleted documentation
Documentation was already added by @spacewolfXfr in recent commits, so this older version of documentation has been deleted in case there are any conflicts and to avoid any confusion. There might be some remnants of it left somewhere in the folder structure, so if this is found in the future, it should be deleted.